requirements.txt
asgiref==3.11.1
Django==6.0.4
dotenv==0.9.9
pyactiveresource==2.2.2
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
ShopifyAPI==12.7.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
